chore: remove unused code
- move shared styling & components to src/shared/Styles.tsx
- remove unused static graphs
- remove old graph-file list handling
- remove unused translation handling
- remove GraphManagerContainer (removes duplicate graph data call to
  backend!)
