--- src/GraphManager/components/index.ts
@@ -1,3 +1,0 @@
-export { GraphFileList } from "./GraphFileList";
-export { GraphManagementMenu } from "./GraphManagementMenu";
-export { VoteDialog } from "./VoteDialog";
