--- src/GraphManager/utilities/getTranslation.tsx
@@ -1,11 +1,0 @@
-import { Text } from "../hooks/types";
-
-export const getTranslation = ({
-  translatedField,
-  language,
-}: {
-  translatedField: Text;
-  language: string;
-}): string =>
-  translatedField.translations.find((entry) => entry.language === language)
-    ?.content || "Translation not found!";
