--- src/graphdata/crypto-1.json.notes
@@ -1,1 +1,0 @@
-only learning connections; math part might be to abstract
