--- src/graphdata/crypto-1b.json.notes
@@ -1,1 +1,0 @@
-only learning connections; math part less abstract
