--- src/graphdata/math-1.ts
@@ -1,118 +1,0 @@
-import { BackendGraphData } from "GraphManager/types";
-// prettier-ignore
-const data: BackendGraphData = {
-  nodes: [
-    { id: "arithm-ops", description: "Basic Arithmetic Operations" },
-    { id: "Addition", description: "Addition" },
-    { id: "Subtraction", description: "Subtraction" },
-    { id: "Counting", description: "Counting" },
-    { id: "Numbers", description: "Numbers" },
-    { id: "Sums", description: "Sums" },
-    { id: "Product of Sequence", description: "Product of Sequence" },
-    { id: "Conditionals (math)", description: "Conditionals (math)" },
-    { id: "Multiplication", description: "Multiplication" },
-    { id: "Division", description: "Division" },
-    { id: "Modulo", description: "Modulo" },
-    { id: "Factorial", description: "Factorial" },
-    { id: "Prime Numbers", description: "Prime Numbers" },
-    { id: "Probability", description: "Probability" },
-    { id: "Conditional Probability", description: "Conditional Probability"},
-    { id: "Exponentiation", description: "Exponentiation" },
-    { id: "Logarithm", description: "Logarithm" },
-    { id: "Equation", description: "Equation" },
-    { id: "DifferentialEquation", description: "Differential Equation"},
-    { id: "Derivative", description: "Derivative" },
-    { id: "Functional Mathematics", description: "Functional Mathematics"},
-    { id: "Limits", description: "Limits" },
-    { id: "Functions (math)", description: "Functions (math)" },
-    { id: "Variables (math)", description: "Variables (math)" },
-    { id: "Topological Space", description: "Topological Space" },
-    { id: "neighbourhood (math)", description: "neighbourhood (math)"},
-    { id: "Fitting (linear)", description: "Fitting (linear)" },
-    { id: "Analysis", description: "Analysis" },
-    { id: "LinearAlgebra", description: "Linear Algebra" },
-    { id: "Calculus", description: "Calculus" },
-    { id: "DiffGeom", description: "Differential Geometry" },
-  ],
-  links: [
-    { id: "01", source: "Addition", target: "Counting", value: 10 },
-    { id: "b1", source: "Addition", target: "Numbers", value: 10 },
-    { id: "02", source: "Counting", target: "Numbers", value: 10 },
-    { id: "02", source: "Multiplication", target: "Addition", value: 10 },
-    { id: "03", source: "Sums", target: "Addition", value: 10 },
-    { id: "b3", source: "Sums", target: "Conditionals (math)", value: 10 },
-    { id: "04", source: "Product of Sequence", target: "Conditionals (math)", value: 10 },
-    { id: "05", source: "Product of Sequence", target: "Sums", value: 2 },
-    { id: "06", source: "Product of Sequence", target: "Multiplication", value: 10 },
-    { id: "07", source: "Division", target: "Multiplication", value: 10 },
-    { id: "08", source: "Modulo", target: "Division", value: 10 },
-    { id: "09", source: "Division", target: "Modulo", value: 10 },
-    { id: "10", source: "Prime Numbers", target: "Division", value: 10 },
-    { id: "11", source: "Prime Numbers", target: "Numbers", value: 10 },
-    { id: "12", source: "Factorial", target: "Multiplication", value: 10 },
-    { id: "13", source: "Factorial", target: "Product of Sequence", value: 4 },
-    { id: "14", source: "Factorial", target: "Addition", value: 10 },
-    { id: "15", source: "Probability", target: "Division", value: 10 },
-    { id: "16", source: "Probability", target: "Multiplication", value: 5 },
-    { id: "b6", source: "Probability", target: "Analysis", value: 5 },
-    { id: "17", source: "Conditional Probability", target: "Probability", value: 10 },
-    { id: "18", source: "Conditional Probability", target: "Sums", value: 2 },
-    { id: "19", source: "Conditional Probability", target: "Logarithm", value: 4 },
-    { id: "20", source: "Conditional Probability", target: "Equation", value: 7 },
-    { id: "21", source: "Exponentiation", target: "Multiplication", value: 10 },
-    { id: "22", source: "Exponentiation", target: "Division", value: 10 },
-    { id: "23", source: "Exponentiation", target: "Addition", value: 10 },
-    { id: "24", source: "Logarithm", target: "Exponentiation", value: 10 },
-    { id: "25", source: "DifferentialEquation", target: "Equation", value: 10 },
-    { id: "26", source: "DifferentialEquation", target: "Derivative", value: 10 },
-    { id: "27", source: "DifferentialEquation", target: "Functional Mathematics", value: 10 },
-    { id: "28", source: "Functional Mathematics", target: "Equation", value: 10 },
-    { id: "29", source: "Functional Mathematics", target: "DifferentialEquation", value: 2 },
-    { id: "30", source: "Derivative", target: "Limits", value: 10 },
-    { id: "31", source: "Limits", target: "Topological Space", value: 2 },
-    { id: "32", source: "Limits", target: "neighbourhood (math)", value: 2 },
-    { id: "33", source: "Limits", target: "Functions (math)", value: 8 },
-    { id: "34", source: "Functions (math)", target: "Variables (math)", value: 10 },
-    { id: "35", source: "Equation", target: "Variables (math)", value: 10 },
-    { id: "36", source: "Fitting (linear)", target: "Functions (math)", value: 10 },
-    { id: "37", source: "Fitting (linear)", target: "Sums", value: 10, note: "Least Square Regression", },
-    { id: "38", source: "Fitting (linear)", target: "Exponentiation", value: 10, note: "Least Square Regression" },
-    { id: "39", source: "arithm-ops", target: "Addition", value: 10 },
-    { id: "40", source: "arithm-ops", target: "Subtraction", value: 10 },
-    { id: "41", source: "arithm-ops", target: "Multiplication", value: 10 },
-    { id: "42", source: "arithm-ops", target: "Division", value: 10 },
-    { id: "43", source: "Probability", target: "arithm-ops", value: 10 },
-    { id: "44", source: "Exponentiation", target: "arithm-ops", value: 10 },
-    { id: "45", source: "Analysis", target: "Numbers", value: 3 },
-    { id: "46", source: "Analysis", target: "Sums", value: 3 },
-    { id: "47", source: "Analysis", target: "Product of Sequence", value: 3 },
-    { id: "48", source: "Analysis", target: "Limits", value: 3 },
-    { id: "49", source: "Analysis", target: "Functions (math)", value: 6 },
-    { id: "50", source: "Analysis", target: "Logarithm", value: 6 },
-    { id: "50", source: "Analysis", target: "Conditionals (math)", value: 6 },
-    { id: "51", source: "Calculus", target: "Limits", value: 10 },
-    { id: "52", source: "Calculus", target: "Analysis", value: 6 },
-    { id: "53", source: "Calculus", target: "LinearAlgebra", value: 6 },
-    { id: "54", source: "Calculus", target: "Derivative", value: 6 },
-    { id: "55", source: "Calculus", target: "Functions (math)", value: 6 },
-    { id: "56", source: "Calculus", target: "Topological Space", value: 6 },
-    { id: "57", source: "DiffGeom", target: "Analysis", value: 6 },
-    { id: "58", source: "DiffGeom", target: "Calculus", value: 6 },
-    { id: "59", source: "Subtraction", target: "Addition", value: 6 },
-    { id: "60", source: "Subtraction", target: "Numbers", value: 6 },
-    { id: "61", source: "Subtraction", target: "Counting", value: 6 },
-    { id: "62", source: "Addition", target: "Numbers", value: 6 },
-  ],
-};
-let idx = 1000;
-data.nodes.forEach((node) => {
-  data.links.push({
-    id: idx.toString(),
-    source: node.id,
-    target: "Mathematics",
-    value: 1,
-  });
-  idx += 1;
-});
-data.nodes.push({ id: "Mathematics", description: "Mathematics" });
-export default data;
