--- src/graphdata/math-simple.json.notes
@@ -1,1 +1,0 @@
-basic math dependencies
