--- src/graphdata/startup.ts
@@ -1,110 +1,0 @@
-import { BackendGraphData } from "GraphManager/types";
-const data: BackendGraphData = {
-  nodes: [
-    { id: "Cetup.ID", description: "Cetup.ID" },
-    { id: "Motivation", description: "Why found a startup?", group: 1 },
-    { id: "Understand Problem", description: "Understand Problem", group: 1 },
-    {
-      id: "Diffusion of Innovation Model",
-      description: "Diffusion of Innovation Model",
-      group: 1,
-    },
-    { id: "Early Adopters", description: "Early Adopters", group: 2 },
-    { id: "Buying Center", description: "Buying Center", group: 2 },
-    { id: "Interview planning", description: "Interview planning", group: 2 },
-    { id: "NABC-Model", description: "NABC-Model", group: 2 },
-    { id: "Need", description: "Need", group: 2 },
-    { id: "Approach", description: "Approach", group: 2 },
-    { id: "Benefit", description: "Benefit", group: 2 },
-    { id: "Competition", description: "Competition", group: 2 },
-    { id: "JTBD", description: "Jobs to be Done", group: 2 },
-    { id: "Pitching", description: "Pitching", group: 2 },
-    { id: "Hypothesis Check", description: "Hypothesis Check", group: 2 },
-    {
-      id: "Competition Analysis",
-      description: "Competition Analysis",
-      group: 2,
-    },
-    { id: "USP", description: "Unique Selling Proposition", group: 2 },
-    { id: "VPC", description: "Value Proposition Canvas", group: 2 },
-    {
-      id: "Problem-Solution-Fit",
-      description: "Problem-Solution-Fit",
-      group: 2,
-    },
-    { id: "Unfair Advantage", description: "Unfair Advantage", group: 2 },
-    { id: "ERKS", description: "ERKS Framework", group: 2 },
-    { id: "Business Model", description: "Business Model", group: 2 },
-    {
-      id: "Business Model Canvas",
-      description: "Business Model Canvas",
-      group: 2,
-    },
-    { id: "Lean Canvas", description: "Lean Canvas", group: 2 },
-    { id: "Storytelling", description: "Storytelling", group: 2 },
-    { id: "Final Pitch", description: "Final Pitch", group: 2 },
-    { id: "Intention", description: "Intention", group: 2 },
-    { id: "Story", description: "Story", group: 2 },
-    { id: "Team", description: "Team", group: 2 },
-    { id: "Research", description: "Research", group: 2 },
-    { id: "Learnings", description: "Learnings", group: 2 },
-    { id: "Revenue Model", description: "Revenue Model", group: 2 },
-    { id: "Pivot/Persevere", description: "Pivot/Persevere", group: 2 },
-  ],
-  links: [
-    {
-      id: "1",
-      source: "Diffusion of Innovation Model",
-      target: "Early Adopters",
-      value: 1,
-    },
-    { id: "2", source: "NABC-Model", target: "Need", value: 1 },
-    { id: "3", source: "NABC-Model", target: "Approach", value: 0.5 },
-    { id: "4", source: "NABC-Model", target: "Benefit", value: 0.5 },
-    { id: "5", source: "NABC-Model", target: "Competition", value: 0.5 },
-    { id: "5", source: "Cetup.ID", target: "Motivation", value: 0.5 },
-    { id: "5", source: "Cetup.ID", target: "Understand Problem", value: 0.5 },
-    {
-      id: "5",
-      source: "Cetup.ID",
-      target: "Diffusion of Innovation Model",
-      value: 0.5,
-    },
-    { id: "5", source: "Cetup.ID", target: "Buying Center", value: 0.5 },
-    { id: "5", source: "Cetup.ID", target: "Interview planning", value: 0.5 },
-    { id: "5", source: "Cetup.ID", target: "NABC-Model", value: 0.5 },
-    { id: "5", source: "Cetup.ID", target: "Pitching", value: 1 },
-    { id: "5", source: "Cetup.ID", target: "JTBD", value: 0.5 },
-    { id: "5", source: "Cetup.ID", target: "Hypothesis Check", value: 0.5 },
-    { id: "5", source: "Cetup.ID", target: "Competition Analysis", value: 0.5 },
-    { id: "5", source: "Cetup.ID", target: "USP", value: 0.5 },
-    { id: "5", source: "USP", target: "Problem-Solution-Fit", value: 1 },
-    { id: "5", source: "USP", target: "VPC", value: 1 },
-    { id: "5", source: "USP", target: "ERKS", value: 1 },
-    { id: "5", source: "Cetup.ID", target: "Business Model", value: 0.5 },
-    {
-      id: "5",
-      source: "Business Model",
-      target: "Business Model Canvas",
-      value: 0.5,
-    },
-    { id: "5", source: "Business Model", target: "Lean Canvas", value: 0.5 },
-    { id: "5", source: "Business Model", target: "Business Model", value: 0.5 },
-    { id: "5", source: "Pitching", target: "Storytelling", value: 0.5 },
-    { id: "5", source: "Final Pitch", target: "Storytelling", value: 0.5 },
-    { id: "5", source: "Pitching", target: "Storytelling", value: 0.5 },
-    { id: "5", source: "Final Pitch", target: "Pitching", value: 0.5 },
-    { id: "5", source: "Final Pitch", target: "Pitching", value: 0.5 },
-    { id: "5", source: "Story", target: "Storytelling", value: 0.5 },
-    { id: "5", source: "Final Pitch", target: "Story", value: 0.5 },
-    { id: "5", source: "Final Pitch", target: "Intention", value: 0.5 },
-    { id: "5", source: "Final Pitch", target: "Research", value: 0.5 },
-    { id: "5", source: "Final Pitch", target: "Learnings", value: 0.5 },
-    { id: "5", source: "Final Pitch", target: "Revenue Model", value: 0.5 },
-    { id: "5", source: "Final Pitch", target: "Pivot/Persevere", value: 0.5 },
-    { id: "5", source: "Final Pitch", target: "Team", value: 0.5 },
-    { id: "5", source: "USP", target: "Unfair Advantage", value: 0.5 },
-    { id: "5", source: "Cetup.ID", target: "Final Pitch", value: 0.5 },
-  ],
-};
-export default data;
